Move dask.async to dask.scheduler
In Python 3.7 `async` will become a keyword and `dask.async` will no
longer be a valid module name. We move `dask.async` to `dask.scheduler`
and deprecate the old module. The documentation has been updated
accordingly.
