+++ dask/dataframe/io/tests/test_hdf.py
@@ -313,7 +313,7 @@ def test_to_hdf_exceptions():
                 a.to_hdf(hdf, '/data_*_*')
 
 
-@pytest.mark.parametrize('get', [dask.async.get_sync,
+@pytest.mark.parametrize('get', [dask.scheduler.get_sync,
                                  dask.threaded.get,
                                  dask.multiprocessing.get])
 @pytest.mark.parametrize('npartitions', [1, 4, 10])
