+++ dask/dataframe/tests/test_hyperloglog.py
@@ -49,7 +49,7 @@ rs = np.random.RandomState(96)
 def test_basic(df, npartitions):
     ddf = dd.from_pandas(df, npartitions=npartitions)
 
-    approx = ddf.nunique_approx().compute(get=dask.async.get_sync)
+    approx = ddf.nunique_approx().compute(get=dask.scheduler.get_sync)
     exact = len(df.drop_duplicates())
     assert abs(approx - exact) <= 2 or abs(approx - exact) / exact < 0.05
 
@@ -60,7 +60,7 @@ def test_split_every(split_every, npartitions):
     df = pd.Series([1, 2, 3] * 1000)
     ddf = dd.from_pandas(df, npartitions=npartitions)
 
-    approx = ddf.nunique_approx(split_every=split_every).compute(get=dask.async.get_sync)
+    approx = ddf.nunique_approx(split_every=split_every).compute(get=dask.scheduler.get_sync)
     exact = len(df.drop_duplicates())
     assert abs(approx - exact) <= 2 or abs(approx - exact) / exact < 0.05
